requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
fake-useragent==0.1.11
flickrapi==2.4.0
ibm-cloud-sdk-core==1.5.1
ibm-watson==4.5.0
idna==2.9
joblib==0.15.1
lxml==4.5.1
nltk==3.5
oauthlib==3.1.0
pathlib==1.0.1
Pillow==7.1.2
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
regex==2020.5.14
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
tqdm==4.46.1
urllib3==1.25.9
websocket-client==0.48.0
